test: Add test for ClothRepository
- Write tests for ClothRepository Class
- Remove CategorySubCategoryRepository test which is no longer needed
